+++ module/ocr/ocr.py
@@ -21,13 +21,9 @@ class OCR:
             except Exception as e:
                 logger.error(_("初始化OCR失败：{e}").format(e=e))
                 self.ocr = None
-                import cpufeature
-                if not cpufeature.CPUFeature["AVX2"]:
-                    logger.error(_("CPU不支持AVX2指令集"))
-                else:
-                    logger.info(_("请尝试重新下载或解压"))
-                    logger.info(_("若 Win7 报错计算机中丢失 VCOMP140.DLL，请安装 VC运行库"))
-                    logger.info("https://aka.ms/vs/17/release/vc_redist.x64.exe")
+                logger.info(_("请尝试重新下载或解压"))
+                logger.info(_("若 Win7 报错计算机中丢失 VCOMP140.DLL，请安装 VC运行库"))
+                logger.info("https://aka.ms/vs/17/release/vc_redist.x64.exe")
                 input(_("按回车键关闭窗口. . ."))
                 sys.exit(1)